% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_richness_campaigns.R
\name{mod_map_richness_campaigns_server}
\alias{mod_map_richness_campaigns_server}
\title{map_richness_campaigns Server Functions}
\usage{
mod_map_richness_campaigns_server(id, downloaded_site_name)
}
\arguments{
\item{downloaded_site_name}{data.frame of coleo sites. Must have specific
column names: "type" for the habitat type. "richness" for the species
richness and "display_name" for the name you want users to see}
}
\description{
map_richness_campaigns Server Functions
}
